Split registry tests by subsystem
Break the flat skills registry test module into focused discovery, install, lookup, and removal submodules. This keeps the thin router file small, preserves the existing test bodies and attributes, and reduces the low-cohesion signal that CodeScene was reporting for the monolithic test module.
